Adds separate html and sass as well as inline html
For email template. Inline html will be the final product used in
emails. Includes a header, picture, and text as well as
call to action buttons at the bottom. Next need to make
preheader same color as background color it is on so it is invisible
and style to our own personal brand